Turned IConnectable into a capability
This has a small side-effect that the master position is not stored as
JSON anymore, but as native NBT data. This would break existing worlds.

I guess this could be changed back, but an additional improvement that
should be made is adding the masters dimension alongside its BlockPos,
so transdimensional storage networks could work. E.g. when building half
your network inside compact machines.
That would also break existing worlds.

Unless we add a way to auto update from the old data to the new one.
Not sure how to handle that yet.