Make all connectables dimension aware
This is necessary if a network needs to span multiple dimensions, e.g.
when connecting through a compact machine tunnel.

This commit turns the connectables set from a BlockPos to a DimPos set,
which additionally stores the dimension the block is in. It also changes
the addConnectables method to iterate through the positions returned by
the capability itself.

Missing is only the client part. It does not currently recognize that
the compact machine has a connectable on the other side of its tunnel.