Fix Import/Export cables not doing anything
The addition of a argumentless constructor so vanilla can instantiate it
caused the EnumStorageDirection to always be set to BOTH, which is a
state that causes Import/Export cables to do nothing, since they would
start behaving very unpredictable. Race conditions all over the place.

This should really be fixed by further splitting the cables down
into two separate tile entities. What we are doing here is a bit of a
hack.