Fix concurrent modification exception by manually synchronizing
The refreshNetwork method is getting called by individual blocks when
they are being loaded. This caused CMEs when the set the refreshNetwork
method is rebuilding is currently being iterated by the master.

To solve this the refreshNetwork method is now only queueing the task to
refresh the network so the master can handle it on its own and in its own
update tick. This ensures that no one else but the master can directly
modify the set.

Fixes #135